+++ src/components/conversation/comment/votebuttons.tsx
@@ -1,25 +1,22 @@
 import "../../../css/components/conversation/comment/votebuttons.css";
-import React from "react";
+import React, { useState } from "react";
 import { ArrowDropDown, ArrowDropUp } from "@mui/icons-material";
 import { Button, ButtonGroup, Typography } from "@mui/material";
 import { useNotification, useUser } from "../../ContextProvider";
 import { api } from "../../../lib/api";
-import { useThread, useThreadId, useUserVotes } from "../ConversationContext";
+import { useThreadId, useUserVotes } from "../ConversationContext";
 import { parseError } from "../../../lib/parseError";
 import { commentType } from "../../../types/conversation/comment";
 
-export default function VoteButtons(props: { commentId: number }) {
-    const { commentId } = props;
+export default function VoteButtons(props: { comment: commentType }) {
     const threadId = useThreadId();
     const [thread, setThread] = useThread();
     const [votes, setVotes] = useUserVotes();
     const [, setNotification] = useNotification();
     const [user] = useUser();
+    const [comment, setComment] = useState(props.comment);
 
-    const comment = thread?.conversation?.find((i) => i.id === commentId) as commentType;
-    if (!comment) return <React.Fragment />;
-
-    const vote = votes?.[commentId];
+    const vote = votes?.[comment.id];
     const up = comment.U || 0;
     const down = comment.D || 0;
 
@@ -28,23 +25,14 @@ export default function VoteButtons(props: { commentId: number }) {
      * @param {"U" | "D"} newVote - "U" | "D"
      */
     function sendVote(newVote: "U" | "D") {
-        if (thread) {
-            thread.conversation[thread.conversation.findIndex((i) => i.id === commentId)][
-                newVote
-            ] = (comment[newVote] || 0) + 1;
-            setThread(thread);
-        }
-        setVotes({ ...votes, [commentId]: newVote });
         api.threads.comments
-            .vote({ threadId, commentId: commentId, vote: newVote })
+            .vote({ threadId, commentId: comment.id, vote: newVote })
+            .then(() => {
+                // for the refetch effect to work, we need to fetch the comment again
+                setComment({ ...comment, [newVote]: (comment[newVote] || 0) + 1 });
+                setVotes({ ...votes, [comment.id]: newVote });
+            })
             .catch((err) => {
-                setVotes(votes);
-                if (thread) {
-                    thread.conversation[
-                        thread.conversation.findIndex((i) => i.id === commentId)
-                    ] = { ...comment, [newVote]: comment[newVote] };
-                    setThread(thread);
-                }
                 setNotification({
                     open: true,
                     text: parseError(err),