+++ src/components/menu.tsx
@@ -1,12 +1,24 @@
 import "../css/components/menu.css";
-import React, { memo } from "react";
+import React, { memo, useCallback, useMemo } from "react";
 import { Box } from "@mui/material";
-import { useData, useMenu, useSearch, useSelected } from "./MenuProvider";
+import {
+    useReFetch,
+    useMenu,
+    useProfile,
+    useRecall,
+    useSearch,
+    useSelected,
+    useSmode,
+} from "./MenuProvider";
 import { useBack, useQuery, useSettingsOpen } from "./ContextProvider";
 import SearchBar from "./searchbar";
 import { useNavigate } from "react-router-dom";
 import loadable from "@loadable/component";
 import { Add, Autorenew, Settings } from "@mui/icons-material";
+import SwipeableViews from "react-swipeable-views";
+import { virtualize } from "react-swipeable-views-utils";
+
+const VirtualizeSwipeableViews = virtualize(SwipeableViews);
 
 const Dock = loadable(() => import("./dock"));
 const MenuTop = loadable(() => import("./menu/menuTop"));
@@ -14,13 +26,28 @@ const MenuBody = loadable(() => import("./menu/menuBody"));
 
 function Menu() {
     const [selected, setSelected] = useSelected();
-    const [, setData] = useData();
+    const [data, setReFetch] = useReFetch();
     const [menu] = useMenu();
     const [search] = useSearch();
+    const [smode] = useSmode();
+    const [profile] = useProfile();
+    const [recall] = useRecall();
     const [query, setQuery] = useQuery();
     const [, setBack] = useBack();
     const navigate = useNavigate();
     const [, setSettingsOpen] = useSettingsOpen();
+
+    const mode =
+        (search && "search") || (profile && "profile") || (recall && "recall") || "menu";
+
+    const slideRenderer = useCallback(
+        (props: { key: number; index: number }) => {
+            const { index } = props;
+            return <MenuBody key={index} selected={index} />;
+        },
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+        [data, smode]
+    );
 
     return (
         <Box
@@ -34,7 +61,7 @@ function Menu() {
                     {
                         icon: <Autorenew />,
                         action: () => {
-                            setData([]);
+                            setReFetch(true);
                         },
                     },
                     {
@@ -55,12 +82,11 @@ function Menu() {
             {/*latest and viral*/}
             <MenuTop
                 refresh={() => {
-                    setData([]);
+                    setReFetch(true);
                 }}
                 onClick={(e: number) => {
                     if (selected !== e) {
                         setSelected(e);
-                        setData([]);
                     }
                 }}
                 selected={selected}
@@ -77,7 +103,7 @@ function Menu() {
                                 if (e.key === "Enter" && query) {
                                     // navigate with router lib
                                     navigate(`/search?q=${encodeURIComponent(query)}`);
-                                    setData([]);
+                                    setReFetch(true);
                                     setBack(`/search?q=${encodeURIComponent(query)}`);
                                 }
                             }}
@@ -85,7 +111,22 @@ function Menu() {
                     </div>
                 </div>
             )}
-            <MenuBody />
+            {useMemo(
+                () => (
+                    <VirtualizeSwipeableViews
+                        key={mode}
+                        index={selected}
+                        onChangeIndex={(idx) => {
+                            setSelected(idx);
+                        }}
+                        containerStyle={{ flex: 1 }}
+                        slideCount={{ menu: 2, profile: 2, search: 3, recall: 1 }[mode]}
+                        slideRenderer={slideRenderer}
+                        enableMouseEvents={true}
+                    />
+                ),
+                [mode, selected, setSelected, slideRenderer]
+            )}
         </Box>
     );
 }