+++ src/components/textEditor.tsx
@@ -3,13 +3,8 @@ import { Editor } from "@tinymce/tinymce-react";
 import { useIsSmallScreen } from "./ContextProvider";
 import { Box, SxProps, Theme } from "@mui/material";
 import axios from "axios";
+import { parseError } from "../lib/parseError";
 
-/**
- * It creates a text editor that can be used to edit text
- * @param {(a: string, editor: import("tinymce/tinymce").Editor) => void} props.changehandler function triggered on editor change
- * @param {string} props.text initial text of the editor
- * @returns A text editor.
- */
 export default function TextEditor(props: {
     onChange?: (a: string, editor: import("tinymce/tinymce").Editor) => void;
     initText?: string;
@@ -32,7 +27,9 @@ export default function TextEditor(props: {
         noMenuBar,
         noStatusBar,
     } = props;
+
     const isSmallScreen = useIsSmallScreen();
+
     return (
         <Box sx={sx} className={className}>
             <Editor
@@ -47,10 +44,92 @@ export default function TextEditor(props: {
                         "https://cdn.jsdelivr.net/npm/metahkg-css@1.0.9/dist/tinymce/skins/content/metahkg-dark/content.min.css",
                     branding: false,
                     ...(noStatusBar && { statusbar: false }),
+                    setup: (editor) => {
+                        editor.ui.registry.addButton("uploadimage", {
+                            icon: "upload",
+                            onAction: () => {
+                                const input = document.createElement("input");
+                                input.type = "file";
+                                input.accept = "image/*";
+                                input.onchange = async (e) => {
+                                    const file = (e.target as HTMLInputElement)
+                                        ?.files?.[0];
+                                    if (file) {
+                                        editor.windowManager.open({
+                                            title: "Upload image",
+                                            body: {
+                                                type: "panel",
+                                                items: [
+                                                    {
+                                                        type: "htmlpanel",
+                                                        html: "<p>Uploading your image...</p>",
+                                                    },
+                                                ],
+                                            },
+                                            buttons: [
+                                                {
+                                                    type: "submit",
+                                                    text: "OK",
+                                                    buttonType: "primary",
+                                                },
+                                            ],
+                                            onSubmit: () => {
+                                                editor.windowManager.close();
+                                            },
+                                        });
+                                        const formData = new FormData();
+                                        formData.append("image", file);
+                                        axios
+                                            .post("https://api.na.cx/upload", formData, {
+                                                headers: {
+                                                    "Content-Type": "multipart/form-data",
+                                                },
+                                            })
+                                            .then((res) => {
+                                                editor.windowManager.close();
+                                                editor.insertContent(
+                                                    `<img alt="" src="${res.data.url}" />`
+                                                );
+                                            })
+                                            .catch((err) => {
+                                                editor.windowManager.close();
+                                                editor.windowManager.open({
+                                                    title: "Error",
+                                                    body: {
+                                                        type: "panel",
+                                                        items: [
+                                                            {
+                                                                type: "alertbanner",
+                                                                text: `Error uploading image: ${parseError(
+                                                                    err
+                                                                )}`,
+                                                                level: "error",
+                                                                icon: "warning",
+                                                            },
+                                                        ],
+                                                    },
+                                                    buttons: [
+                                                        {
+                                                            type: "submit",
+                                                            text: "OK",
+                                                            buttonType: "primary",
+                                                        },
+                                                    ],
+                                                    onSubmit: () => {
+                                                        editor.windowManager.close();
+                                                    },
+                                                });
+                                            });
+                                    }
+                                };
+                                input.click();
+                            },
+                        });
+                    },
                     mobile: {
                         menubar: noMenuBar ? false : "file edit view insert format tools",
                         toolbar:
-                            "undo redo | bold italic underline strikethrough | image template link codesample | fontsize blocks | alignleft aligncenter alignright alignjustify | outdent indent |  numlist bullist | forecolor backcolor removeformat | pagebreak | charmap emoticons | fullscreen preview save print | insertfile media anchor | ltr rtl",
+                            "undo redo | bold italic underline strikethrough | emoticons | uploadimage image template link codesample | fontsize blocks | alignleft aligncenter alignright alignjustify | outdent indent |  numlist bullist | forecolor backcolor removeformat | pagebreak | charmap | fullscreen preview save print | insertfile media anchor | ltr rtl",
                     },
                     quickbars_selection_toolbar:
                         "cut copy paste | formatselect | quicklink",
@@ -62,7 +141,7 @@ export default function TextEditor(props: {
                         autoresize ? "autoresize" : ""
                     } preview importcss searchreplace autolink autosave save directionality code visualblocks visualchars fullscreen image link media template codesample table charmap pagebreak nonbreaking anchor insertdatetime advlist lists wordcount help charmap quickbars emoticons`,
                     toolbar:
-                        "undo redo | bold italic underline strikethrough | image template link codesample | fontsize blocks | alignleft aligncenter alignright alignjustify | outdent indent |  numlist bullist | forecolor backcolor removeformat | pagebreak | charmap emoticons | fullscreen preview save print | insertfile media anchor | ltr rtl",
+                        "undo redo | bold italic underline strikethrough | emoticons | uploadimage image template link codesample | fontsize blocks | alignleft aligncenter alignright alignjustify | outdent indent |  numlist bullist | forecolor backcolor removeformat | pagebreak | charmap | fullscreen preview save print | insertfile media anchor | ltr rtl",
                     toolbar_sticky: toolbarSticky,
                     toolbar_mode: "sliding",
                     ...(toolbarBottom && { toolbar_location: "bottom" }),
@@ -97,7 +176,7 @@ export default function TextEditor(props: {
                     autosave_restore_when_empty: false,
                     autosave_retention: "2m",
                     image_advtab: true,
-                    images_upload_handler: async (blobInfo, progress) => {
+                    images_upload_handler: async (blobInfo, _progress) => {
                         const formData = new FormData();
                         formData.append("image", blobInfo.blob());
                         const { data } = await axios.post(